backend/requirements.txt
# FastAPI
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.31.0
alembic>=1.14.1

# Redis
redis[hiredis]>=5.2.1

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.1

# AI
openai>=1.58.1
deepl>=1.18.0
langdetect>=1.0.9

# HTTP
httpx>=0.28.1

# Config
pydantic-settings>=2.7.1

# Utilities
python-dateutil>=2.9.0
